requirements.txt
python-dotenv
emoji
